+++ tractor/ipc/__init__.py
@@ -1,0 +1,11 @@
+from ._chan import (
+    _connect_chan,
+    MsgTransport,
+    Channel
+)
+
+__all__ = [
+    '_connect_chan',
+    'MsgTransport',
+    'Channel'
+]
